better parallelization; clean up main program
